cache: Add use_trafficserver_tls parameter to unified profile
Allow signaling the unified profile that it should use ATS as the TLS
termination layer instead of nginx. Right now this is used to disable
the redir_port in tlsproxy::localssl. Take into account that
http -> https redirection is actually being performed by varnish-fe
so this doesn't have any impact on production traffic.

Bug: T221594
Change-Id: I8abcb77b418ea3ec169142b518743df63034b6a9