RESTBase: Temporarily allow access to port 7233 as well
Until RESTRouter becomes the default REST API router, we need to expose
the current RESTBase instance on both 7231 and 7233.

Note that it is safe to reuse the `restbase_web` ferm rule because the
basic firewall rule opened for the defined port (7231 in this case) is
already set up by service::node. In other words, `retbase_web` was
useless (until now, that is).

Bug: T223953
Change-Id: I41f1aa21aa47fc6588286e4da71929cf657e2afe