tools-prometheus: add an allowance for ssh monitoring
The firewalled portions of toolforge that used ferm instead of security
groups all show as "ssh unreachable" in our dashboards.  This causes
us to ignore that dashboard, which is pretty unhealthy.  Since it also
clear that this all needs to be refactored soon, I went for a gentle approach
to fix this.  Accepting that the refactor would make this unreadable, I have
simply applied a ferm::service to the appropriate profiles and created a new
one to include in things that are in weird roles.  I used a hiera
value that already exists in toolforge to identify the prometheus servers.
It might need to be added in toolsbeta, but that's ok.

Change-Id: I23459887aa59818b8170e7757b5ac67e0bd264cd