misc_apps::httpd: do not load SSL httpd module
We now want to use envoy for TLS termination and it has been installed
by I97c009926e771.

But httpd listens on 443 because the default ports.conf has a "Listen 443"
if SSL/TLS modules are loaded.

Instead of puppetizing the entire ports.conf as we did in some cases in the
past just stop loading the SSL module and that should do the same because
of the "IfModule" parts in the default ports.conf.

Bug: T224247
Bug: T210411
Change-Id: Ib8e69f8203c46ae97902d3ba402a4d118f78a848