redirects.dat: secure external redirects
All of these external redirects were lacking https:// as the
protocol.  A quick audit showed almost all of them now support TLS
on the same domain.

The only odd case was the "wiki.media.hu" target, which has broken
TLS, but the HTTP listener (that we were redirecting to) was doing
a redirect to https://wikimedia.hu/ , which does have working TLS,
so I've skipped the intermediate and set it go directly there.

Change-Id: I5cbd417ba328d48059d37e429661a25bf6ee96eb