powerdns: correct some database variables in my.cnf
The socket seems to be getting "cleaned up" since it is in /tmp and other
unexpected things could be wrong since it thinks it has the wrong basedir.

Switching to the module's default socket and keying basedir off the OS.
It is awfully difficult to dig deeply into database issues without the socket.

Bug: T224828
Change-Id: Ibf3b4725ad6199b07c4c4fa536142e0f36f6e3a8