mediawiki::php: only enable tideways/mongodb where needed
While neither extensions is a problem per-se, and tideways doesn't
really create issues for us in terms of performance, it seemed to be one
of the possible causes of global contention on a server that was
experiencing large lags. Given we only run profiling via tideways on the
mwdebug hosts, reduce the installation to those.

Change-Id: If5d9b6109367368797d1d1ad5eb2f1e52c9c7fbb