debconf::set: use set-selections instead of communicate
From my tests debconf-communicate seems to be able to set values
for questions only when they are already set by somethingelse
(like package install I suppose), meanwhile debconf-set-selections
seems not to have this restriction.

I am fairly ignorant about debconf but it seems easier to preseed
questions that we already know before the package is installed by puppet
via set-selections, rather than using communicate.

The downside of set-selections seems to be the fact that a type
for the question is needed, meanwhile communicate doesn't have this
requirement (seems consistent with what said above).

Change-Id: Ie9717fdbf2e1af134472c7f8f8d3be347db36897