Generalize Piwik's db backup profile for Analytics
The profile::piwik::backup class has been generalized
to be used for other Analytics Database environments,
most notably the Meta instance (an-coord1001 currently).

This change should be a no-op for Piwik's settings
(except for the Puppet private repo's changes needed), but
it is the groundwork to add periodical backups to
the Analytics meta database.

Note that the backup::mysqlset already takes care of dumping
everydatabase into a single gzipped file.

Bug: T231208
Change-Id: I5358238ba2f36caeb7df0e7ba6e037f332122f99