Phabricator: Make a separate hiera option to ensure phd stopped/running
Currently we do not run phd on backup phab servers. We now want to be
able to configure this separately so that the backup server can potentially
run phd for testing and for warm-spare / failover.

This introduces a separate setting for phd and provides the setting
in relevant host-specific hiera yaml files.

Bug: T232883
Change-Id: Idfb6a5f9e4fdaa0f5284216ed85ed0c687ccd120