profile::mediawiki::webserver: ensure cron_splay is called correctly
cron_splay requires the current host to be a member of the array passed
in the first argument, `$pool_nodes` in this case.  This CR ensures that
is the case.

Currently mwdebug* servers are failing to successfully complete a puppet
run do to this issue.

This issues was introduced in the following change:
  Ibd7d5aff91444e0e5b98ecefd4e52448dcd4aefc

Change-Id: Ie49a1f9ea74090906a08d69889fef8a13c64993f