toolforge-kubernetes: restructure pod security policies
This puts the base installation PSPs into a single file along with adding
the hostpath restrictions to satisfy one of the requirements of PSP.

This has been tested to block both containers and pods running as root,
unless the privileged psp is applied.

We will have to define toolforge tool PSPs in maintain-kubeusers.
Testing that now.

Bug: T227290
Change-Id: I053f0c0ff2996d1017f8524b4b4e8011f87fbd39