backups: Apply no-srv-format recipe to backup hosts
the db-no-srv-format was initially a recipe to allow reimage
without touching /srv. This later was changed into a recipe
that fails to install anyway, as formating the / partition
would be as badly impacting. Rename it to no-srv-format so
it can be applied as well to backup hosts, not only dbs.

That way, if the databases are mistakenly reimaged, they will
fail to do so (preventing data loss) unless they are applied
the proper recipe (raid1-lvm-ext4-srv-plus-hwraid.cfg).

Bug: T229209
Change-Id: I58d7c4abd0e14b94666292dc37ba53bb228d40dc