mariadb: Promote db1123 to s3 master
db1123 will be the new master for s3.
It was going to be db1078, but given that it is in a batch which is
experimenting BBU failures, let's promote db1123 instead.

db1075 will be the official candidate master, it will get its BBU replaced
soon.
Meanwhile, I will leave db1078 with STATEMENT, just in case there is a
failure on db1123 and db1075 didn't get its BBU replaced yet (should be
a matter of days).
Once db1075 is back with a new BBU, we can change its format to
STATEMENT and make it the official candidate master

Bug: T230783
Bug: T233569
Change-Id: I3886515bf69b4cb05757529b83999ebb3d16943b