profile: add mmutf8fix to kafka output actions
one of the core issues behind the logging pipeline thrashing after network outages and the like is
due to non-utf8 charsets entering the pipeline. logstash does not like non-utf8 logs in its utf8
pipeline and these messages can crash the mutate plugin. this proposal is to mitigate the issue at
the shipper level.

Bug: T233662
Change-Id: I88d5b3316717671337953eb4782d26783db19ea9