Remove tmpreaper from mediawiki servers
This was primarily used for HHVM dumping core and we don't need to pull
it in when the mw servers get reimaged following the HHVM removal.

tmpreaper has a number of technical issues (e.g. the incompability
with PrivateTmp on systemd) and for core dumps we now also have
systemd::coredump.

Bug: T151304
Change-Id: I6f0789dafac84b4c418f9ae614c17d354ea32942