openstack: newton: neutron: introduce WMF patches
Introduce modifications to config.py and router_info.py to support our custom
options dmz_cidr and routing_source_ip, which we use in CloudVPS for our
networking architecture.

This patch contains the actual code diff between the original neutron code and
our customization.

Even thought the patch is copy/pasted from the Mitaka version, some small patch
fuzzing was required and some minor updates added to account for the updates
upstream did to the neutron code. I mean, our custom patch is the same as it
was in Mitaka, just refreshed for Newton.

The old Mitaka patch can be seen with the following commands:

 diff -u modules/openstack/files/mitaka/neutron/l3/original/router_info.original \
         modules/openstack/files/mitaka/neutron/l3/router_info.py

 diff -u modules/openstack/files/mitaka/neutron/l3/original/config.original \
         modules/openstack/files/mitaka/neutron/l3/config.py

The diff contained in this very patch can be generated anytime with the
following commands:

 diff -u modules/openstack/files/newton/neutron/l3/original/router_info.py.original \
         modules/openstack/files/newton/neutron/l3/router_info.py

 diff -u modules/openstack/files/newton/neutron/l3/original/config.py.original \
         modules/openstack/files/newton/neutron/l3/config.py

Bug: T233665
Change-Id: I4b2bc8158c47549d942aeadef5ee3b6b68479aa4
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>