openstack: neutron: newton: refresh original files
These are the original config.py and router_info.py file that are shipped with
openstack Newton, as seen in the Debian package python-neutron
in version v2:9.1.1-3+deb9u1.

The original files in the .deb package tree are:
* /usr/lib/python2.7/dist-packages/neutron/conf/agent/l3/config.py
* /usr/lib/python2.7/dist-packages/neutron/agent/l3/router_info.py

This patch updates also the "patched" files to be a copy of the original.
Starting from here, we will work on patching the final files that will be
deployed via puppet into our cluster.

One can download the original .deb package from:
<http://snapshot.debian.org/package/neutron/2%3A9.1.1-3%2Bdeb9u1/#python-neutron_2:3a:9.1.1-3:2b:deb9u1>

We store the original upstream files in the puppet tree to track our local
patches for Neutron.
While at it, rename files to introduce some consistency in the name pattern.

Bug: T233665
Change-Id: I98e082e9c1dc1d202f78eccacb97a92c2e15de6b
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>