mariadb::packages: support buster, drop libmariadbclient18
On a buster cloud instance using this class:

Package libmariadbclient18 is not available, but is referred to by another package.
This may mean that the package is missing, has been obsoleted, or
is only available from another source

However the following packages replace it:
  libmariadb3

Though this should not have to be listed here at all because the -client and -server
packages pull it in.

Change-Id: If9a1b8cf295a394812e5cfc6fc71467a0aebcbba