gerrit: G1GC tuning. Increase NewGen space
The heap usage on Gerrit is only about 10GB, but the garbage collector
keeps getting triggered despite there being 20GB allocated to the heap.
This is because there is a lot of churn in the NewGen space due to bots
crawling various projects.

By default, NewGen space is 5%: 1GB on a 20GB heap. This matches up with
the current allocations hashar found looking at "jmap -heap". This
change increases the NewGen space to start at 3GB and it also removes
the pausetime target (so now it defaults to 200ms). 300ms pause time
targets are a bit too slow for us.

Change-Id: Ie1d27d318a9b97e595ceef6a2d173a6a152543aa