ores: Alert on aberrant amount of celery workers
Arbitrarily define that -2/+2 from the defined $celery_workers is an ok
range for the count of celery processes to be at and alert on processes
being outside of that range. Be very conservative in the alerting as
during deployments, which can be a lengthy process for ORES, there will
be big fluctuations of this

This code is in profile::ores::web because that's were the variable
$celery_workers is. Normally, it would be under profile::ores::workers,
but that would be counter intuitive.

Bug: T230917
Change-Id: I981657bab1ba517c120686b13b934d586c8ddef1