parsoid: turn wtp1025 into eqiad parsoid/php appserver
After I5ef55276d06d has been merged and the cert situation
has been straightened out we can pick the first wtp server
in eqiad to be the equivalent to wtp2001.

Now it should simply be this Hiera change that is needed.

One server per DC has been requested for the initial benchmarking.

Bug: T233654
Change-Id: Ib6e601b54465c6dc007a6ce1fd74f91537585bb7