profile::kerberos::kdc: add support for bacula backups
The /srv/backup directory contains dumps of the Kerberos database.
Add the support for bacula backups to each KDC server.

The size of the dumps is in the order of some KBs, for multiple
files (currently ~250). If the number of files in the dir is a problem
they can be trimmed down.

Bug: T226089
Change-Id: Ibc7a8194619700b56d134358bb50542c01e0dd3a