cassandra config updates for 3.11.4 upgrade
cassandra.yaml
--------------
Most all of these changes are non-normative; Most of these are changes
to comments, or to the sequencing of directives.  This is done to
maintain a reasonable signal to noise ratio when comparing our
configuration to the corresponding upstream version.

Where new configuration options are introduced, the values included
here represent the defaults (i.e. they represent what you'd get if they
were omitted).  There is one exception to this, the newly introduced
`enable_materialized_views`, has been set `false` (defaults `true`).

We do not use materialized views.  If that were to change, it should
be a conscious decision made after evaluating the trade-offs; Setting
this to false better documents what our default policy should be with
respect to materialized views.

CASSANDRA-11839 implemented a more robust means of establishing the
health of a stream session.  As a result,
`streaming_socket_timeout_in_ms` has been deprecated in favor of
`streaming_keep_alive_period_in_secs` (where we'll assume the default
for now).  `streaming_socket_timeout_in_ms` is a templated value, but
we need to leave it defined in the Puppet manifest; This is still a
thing for versions 2.2.x.

jvm.options
-----------
The JVM flag -XX:+UseNUMA is now passed by default in 3.11.4.

Since we no longer use cmcd on 3.x in production, pass JVM flag
-XX:+PerfDisableSharedMem per upstream defaults.

logback.xml
-----------
Removes the shutdown hook (uneeded). Update format of timestamps sent
to stdout to match upstream.

Bug: T200803
Change-Id: Ibc444c2bba13c928816f1de73b154b656a3a115a