nova: try to reduce the number of db connections
Our database setup limits us to 100 connections, which is
generally regarded as 'not nearly enough' by most openstack
deployers.

This patcj makes two reductions:  It changes the pool size
from 4 to 3, and also changes the number of workers
(of all sorts: compute, api, and conductor) from 8 to 6.

This should reduce our total number of db connections by
more than 40%.  If we /still/ run out of connections
after a while then I suspect a leak.  If we don't run
out of connections we may want to fine-tune these
numbers further.

Bug: T234876
Change-Id: I9db9f53817990cf508f771c61ea2564943ba8ca8