bacula: Add conditional storage device setup
For a temporary period, add a conditional so we keep the old
device setup arrangement for hosts with stretch or lower, while
we apply a new storage device setup for the new hosts, for
buster and higher.

This is only a temporary arrangement until the old hosts are
decommissioned, even if it is a bit ugly to maintain- but
it should work for the few days in which both setups coexist.

Bug: T229209
Change-Id: I3d624aad5516dc2b3b6f92f74b2dbd41994c216b