Remove unused/unnecessary passwords::postgres include
passwords::postgres is only useful/used by the role::postgres::master
and role::postgres::slave OSM roles (in the past these were in production
as labsdb1006/1007 and now running in Cloud VPS), I suppose it was
a copy&paste error when the puppetdb role was created.

Remove it to reduce confusion.

Change-Id: I4b2c3763b275ed070a2dab006c27a39850a93aa2