yarn: Add sequential scheduler queue for heavy jobs
Adds a queue that goes one step beyond the nice queue
and only executes one application at a time. This
queue is intended for particularly heavy jobs that might
be run concurrently by a scheduler (oozie). Mjolnir
runs some jobs per-wiki and will use this to limit it's
aggregate resource usage on a particularly expensive
step.

Change-Id: I3e1e5652d8c66bc5a0e996c3e517c7e00a84ea9e