vcl: support new ACL bot_blocked_nets
There's already blocked_nets, which is an IP address ACL defined in
puppet private for which we'll return a 403 for all traffic.  This
is useful in 'break glass' scenarios, but is a pretty heavy hammer.

So let's add bot_blocked_nets, an IP address ACL that will also be
defined in puppet private, but for which we'll only block traffic if
the requests are also using a generic User-Agent.  This is already
disallowed by policy, albeit one that isn't enforced.  (IMO we should
change this -- possibly as a KR in an upcoming quarter? -- and I see
this change as a very small first step.)

Part of the idea of bot_blocked_nets is it is more appropriate to leave
enabled for a given set of addresses for longer periods of time than
is true of blocked_nets.

Change-Id: Ic5300ff811803f17760a8a42e0d454616ebd37de