No longer use --no-cache when dumping Wikibase entities
This used to disable reading from the empty cache we had
given that this was about the only code running on Zend
while everything else was on hhvm.

I tested this briefly and got between 8 and 15% speedup
when dumping 100,000 randomly picked entities.

Note: This will make Wikibase only read from cache, nothing
is ever written to it, so it is not polluted by us looping
over all entities.

Change-Id: I89eae1e8eba09cc211bbe71ae93a088e250fabd9