openstack: update eqiad1 clients to wikimediacloud auth url
This commit updates all AUTH URLs used by the openstack clients and services
to the new openstack.<region>.wikimediacloud.org FQDN. This change will allow
clients to continue operating when switching the DNS CNAME between controllers.

Now that Keystone is running active/active there should be no dependencies on a single
controller. Previously maintenance and monitoring activities only ran on
cloudcontrol1003. This commit updates these tasks to run on the host which is
currently mapped to the primary endpoint FQDN.

Bug: T223907
Change-Id: Ib5cf04732a3a6ac55e1f93d9a7369b62d5b5f100