toolforge: Remove updatetools script
The updatetools script is used to populate a database of Tool
information which powers some of https://tools.wmflabs.org/admin/. The
script should run as the `tools.admin` user, but systemd has decreed
that `.` is an invalid character in a username [0] (in contradiction to
the POSIX standard) [1]. The easiest fix for this is to move this script
out of Puppet and make it part of the admin tool itself. This has a nice
benefit of making things less confusing for future maintainers.

The removal method used here will *not* clean up existing deployments.
Manual cleanup on tools-sge-services-*.tools.eqiad.wmflabs after this
change has been merged would look something like:
* systemctl stop updatetools
* systemctl disable updatetools
* rm /lib/systemd/system/updatetools.service
* systemctl daemon-reload
* systemctl reset-failed
* rm /usr/local/bin/updatetools

The `profile::toolforge::services::updatetools_enabled` hiera setting in
Horizon can also be deleted as a cleanup step.

[0]:
<https://github.com/systemd/systemd/blob/100d5f6ee6ec53ecc20160cd4d6656cacf57db4f/src/basic/user-util.c#L592-L600>
[1]:
<https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap03.html#tag_03_437>

Bug: T229261
Change-Id: Ib0c6143353c54b635e88bfe683feb47d427094f7