toolforge: exclude grid engine TMPDIR directories from tmpreaper
Adjust /etc/tmpreaper.conf settings to exclude directories matching the
patterns used by Grid Engine as the TMPDIR for grid jobs from being
deleted by tmpreaper. The exclusion pattern should only prevent the
directory itself from being deleted and importantly not the files
contained in those directories. We do want to keep the cleanup of leaked
files that tmpreaper was added for (T190185), but we want to prevent the
removal of the Grid Engine managed directory itself.

Bug: T217815
Change-Id: Ice2b6520fa82f26e35935207d0e98d1aafa074e5