toolforge: refresh puppet code for the new k8s
We introduced the 'kubeadm' prefix into the puppet code to differentiate from
a previous code iteration in which we were trying to use other approach to
build the kubernetes cluster.

Today we have a pretty well established method for building and deploying the
cluster based on kubeadm and we aren't going back to the other approach. We no
longer have the need to maintain two code namespaces (kubeadm or not) since
everything is now based on kubeadm.

Cleanup the code (roles, profiles) we don't use.

Bug: T215531
Change-Id: I94362bab8ae0585ba0a6c6fe6e5459c4a763aaf1
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>