Add Icinga check for monitoring correct application of CPU microcode updates
With Linux 4.19 we noticed a situation where the Linux kernel on three Dell
servers refused to load microcode updates. This could only be remediated with
a firmware update. Add an Icinga check to validate that CPU microcode
updates correctly show up in CPU flags exposed by the Linux kernel.

We have a few servers with CPUs which were not addressed by Intel microcode
updates. These are simply blacklisted in the check, we won't see any new
of these and they'll simply grow out soonish as hardware gets replaced.

This installs the Icinga check, the NRPE config will be deployed in a second
step.

Bug: T235250
Change-Id: Ia23babefbc54392712904470851ac0c7a3e348cc