logstash: increase filter truncate max_counters to 125000
observing warnings in the logstash log to the effect of
"[WARN ][logstash.filters.throttle] filters/LogStash::Filters::Throttle:
Decreased timeslot max_age to 630 because max_counters exceeded by 3898."

since this seems to be only slightly exceeding the default of 100000,
increasing to 125000 should help. if the warnings persists we can restore
the default value and perhaps adjust the key to be less granular.

Change-Id: I59e967b6717c6bf810b02b301a2c5f97bdd72d00