host monitoring: add optional contact group for mgmt interfaces
Since it is entirely possible (and currently desired by WMCS) for one to
want to monitor a host's public interfaces for paging while also wanting a
quieter (non-SMS) alert for the management interface, should allow setting a different
contactgroup in nagios for that interface. Several times WMCS has been
paged when mgmt interfaces were unavailable briefly, and this isn't really
useful. Paging for the main interface is quite useful for WMCS on the same
servers. This requires different contactgroup assignments in nagios.

Bug: T223458
Change-Id: I70da9b30ff5d0e95ba94873b167d55913b80abdb