ATS: Use a common base path for /etc/ssl and /etc/acmecerts certs
ATS lacks the ability of changing the TLS material base paths without
restarting. At the same time we need to be able to change from one
vendor like digicert to Let's Encrypt without being force to restart
ats-tls.

As a side effect it drops support on trafficserver::tls_material
for the old LE puppetization. It hasn't been exposed via the
trafficserver::tls profile, so it shouldn't be a big deal.

This CR will require a restart of ats-tls cluster-wide.

Bug: T234803
Change-Id: I465ac4945a598672a24c7164d5feef1a8c74b111