puppet-merge: implement locking.
This change is another attempt at implementing locking in
puppet-merge[1].  The first attempt failed as the puppet-merge script
calls its self with the labsprivate switch.  As such we need to
maintain two locks.  one for when the script is called with this flag and
one for when it is not.

[1]I2d5b40f3016fd8b159778ba63135212320ef961c

Change-Id: I9c580938ed101b0fd52925e5416ebc0cd8024341