logstash: set template_overwrite true in elasticsearch outputs
Currently changes to the puppet managed logstash elasticsearch template are
not being applied to elasticsearch automatically.

Per the template_overwrite documentation: "if you have your own template file
managed by puppet, for example, and you wanted to be able to update it
regularly, this option could help"

This describes our configuration. We want to apply an updated logstash
template to elasticsearch when it is changed on the filesystem by puppet.
Enabling template_overwrite should address this.

Bug: T234564
Change-Id: I379ab94a3d0c9595346040db9dd14741c57ad76e