bacula: Create new backup jobs status check for icinga
First version of a script to run on the director, that checks
that backups are running correctly- for that it checks, for each
configured job, that there is a fresh, correct backup and a full
backup with the expected date, as well as the last full backup
with non-zero bytes.

It detects and prints information about anomalous states, such
as stale backups, all errors and no backups run. If not backups
are found configured globally, it returns UKNOWN.

Bug: T234900
Change-Id: I4104e2b25ccb15621826dfe0725ab3f69a16d2c0