logstash: remove deprecated elasticsearch options
These settings are removed in ELK 7, causing errors not warnings like now:

[2019-10-22T08:57:50,451][WARN ][logstash.outputs.elasticsearch] You are
using a deprecated config setting "flush_size" set in elasticsearch.
Deprecated settings will continue to work, but are scheduled for removal
from logstash in the future. This setting is no longer necessary as we
now try to restrict bulk requests to sane sizes. See the 'Batch Sizes'
section of the docs.

[2019-10-22T08:57:50,452][WARN ][logstash.outputs.elasticsearch] You are
using a deprecated config setting "idle_flush_time" set in
elasticsearch. Deprecated settings will continue to work, but are
scheduled for removal from logstash in the future. This is a no-op now
as every pipeline batch is flushed synchronously obviating the need for
this option.

Bug: T235891
Change-Id: I3edad6b138a4abffc7dd67ff7ed2f891214291b3