gerrit: change gerrit master_host to gerrit1001, remove duplicate
The master_host setting should only influence rsync setup
and isn't the actual switch to change masters, despite the name.

What actually decides what is the master is "is_replica = false".

But we should do this if we don't need to rsync from cobalt anymore,
it should allow pushing from gerrit1001 to a future host, or 2001.

Remove the duplicate setting of master_host in the same Hiera file.

Bug: T222391
Change-Id: I6c2d53c13fa774ac798df2699ce33eecbc69f322