Varnish: don't decode/encode slashes for core REST API paths
Deciding what canonical encoding to use for MW Core REST APIs
is a larger project tracked under T235782. The plan for Parsoid
REST API outlined in T235478 is not sufficient for the other
core REST APIs, cause the timelines do not match.

Simply not encoding/decoding a forward slash for rest.php routes
should be enough to unblock that feature.

Bug: T235779
Bug: T235478
Change-Id: I7451669ef6cc55dfbb3177b4f45ef5c3580d7a13