gerrit: increase heap_size from 20G to 32G
We now have 64GB of RAM we can use instead of just 32GB.

quote from [1]:

"So what should the default value be? According to Oracle documentation,
starting from J2SE 5.0 it is the smaller of 1/4th of the physical memory
or 1GB (before J2SE 5.0, the default maximum heap size was 64MB).

We recommend you set this parameter to 8GB for a 16GB machine and
16GB or more for a machine with 24GB or more memory."

The first part does not seem to match the actual recommendation, so going
with half of the 64GB.

[] https://collabnet.zendesk.com/hc/en-us/articles/234122407-Memory-settings-in-Gerrit-configuration

Bug: T225166
Bug: T222391
Change-Id: Ibefe4d57a21afccf5f9537bb41ba74cc5dc11f7c