maintain-kubeusers: add ability to merge and update configs
To allow toolforge users to switch clusters, this must not
destroy the other cluster's configuration in the file. Here it uses flock
to make sure it isn't in a fight with the other service, and it will preserve
data it finds in the file, changing only what it must.

This must make a user-writable file so users can change context.

Bug: T236202
Change-Id: I1f4ee973bc7b108c65de0bc09792e6eb0fbaa91b