bacula: Allow nagios user to execute bacula check as bacula user
I forgot npre checks run under the nagios user. Allow it to run
the bacula check, which requires root or, better, bacula
permissions under the bacula user.

This will fix the current error 13 (permission denied) on the
backup freshness alert.

Move also the script from /usr/sbin to /usr/bin as it will be
intended to be run with an unprivileged user.

Bug: T236423
Change-Id: I4ea0a56b7b87bd41a9d66f526af52096c566045f