bacula: Fix check so it produces the right nagios error code
By mistake, despite detecting an error, the check_bacula.py script,
which detects backup freshness, returned always 0 (OK) code.

This was due to a) not creating tests; and b) Jaime handling
several programming languages and not remembering basic primitive
type handling on function argument passing on python.

Fix the argument passing so data is returned and not tried to
mutate unmutable objects.

Bug: T236423
Change-Id: Ia23ce0a775469d9789bd9fa17a2ad91dc64c71c0