puppetmaster: Add feature flag for geoip provisioning
The ::puppetmaster class has historically provisioned MaxMind geoip
databases (either gratis or subscription based) on each puppetmaster to
support the ::geoip::data::puppet class in client manifests. Within the
Cloud VPS environment there is very little need for geoip databases and
the upstream apt package for the gratis version is kept in the "contrib"
component which we do not add to the default apt config.

This patch adds a $enable_geoip feature flag to ::puppetmaster
defaulting to 'true', which is the same as the legacy behavior. It also
adds a $enable_geoip feature flag to ::role::puppetmaster::standalone
which is passed to ::puppetmaster internally. The new feature flag on
::role::puppetmaster::standalone defaults to false, which is new
behavior introduced by this patch. ::role::puppetmaster::standalone is
only intended for use within Cloud VPS projects, so this should not
effect any production hosts.

Instances in Cloud VPS desiring geoip databases can set
`::role::puppetmaster::standalone::enable_geoip: true` in hiera to
restore the previous default behavior of installing MaxMind geoip
databases via apt packages. This may also require adding new apt sources
to the instance which is not done by this patch.

Bug: T236487
Change-Id: Icbf39d57ac23157e2b79eb2f833be1b76e2d7cfd