toolforge-k8s: enable the settings API and PodPreset
This would allow us to mimic the automounter from the custom controllers
in use now. We don't absolutely need it because webservice and some templates
could take care of this. However this would ensure the mounts are present
in any webservice pods or pods with certain labels if added to maintain-kubeusers.

Bug: T215678
Change-Id: I2a862a1f9fc37f01845d5cfc490d7dbd9c4a8eb4