releases: remove php-xdebug
We do not need XDebug on the release hosts since we do not run any PHP
coverage jobs there. The reason the package is installed is that the
profile previously borrowed the list of PHP packages from a CI puppet
class.  When the CI class has been phased out, the list of package got
copy pasted in the releases profile.

The package would have to be manually cleaned up on the hosts as well as
the ini files under /etc/php.

Hosts: releases1001.eqiad.wmnet, releases2001.codfw.wmnet

Bug: T236774
Change-Id: Id94bcb46e6c879671022620ce0dbd98212fad294