zuul: change gear queue monitoring threshold
Zuul triggers functions in the gearman server which then executes it
from two set of workers:
* two zuul-merger on contint1001/contint2001
* the fleet of executors in Jenkins

As I got it the previous alarm looked for 15 minutes of data, and would
raise if 30% of the point were above 140 functions waiting.

Switch to 100% of point over 10 minutes. Slightly raise the number of
functions to 150.

Change-Id: I8f1de4875ddd2160f03a9b5c31db9fef7911b7da