toolforge: k8s: introduce switch for etcd role
We want to reuse the same puppet prefix for the old and new etcd clusters for
kubernetes.
The old cluster is meant to be run in Debian Jessie, while the new one is for
Debian Buster.

The old cluster is using role::toollabs::etcd::k8s while the new one is using
role::wmcs::toolforge::k8s::etcd. With this patch the old role is added in case
the new role is applied to a Debian Jessie VM.

The puppet prefix should be updated to only point to the new role
role::wmcs::toolforge::k8s::etcd.

Hiera shouldn't be a problem since Debian Jessie VMs will still read the old
ones while Debian Buster VMs will read new ones.

Bug: T236826
Change-Id: I967b81d69f177463331089d7994b5ca22f2ef94b
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>