prometheus: Add new prometheus mysqld exporter group: test
test-s1 and test-s4 hosts were in a limbo, they are part of the
s1 and s4 replica set, but they never should be part of
core-production (serve mediawiki traffic). We used to call them
core-test hosts.

However, it was brought to our attention that because of this
duality, aggregated metrics count them twice. This was resolved
already by deleting/updating records on the zarcillo database.

However, they still show up on mw production health metrics-so
I decided to classify them out of core- into a test group on its
own.

This adds such a group to prometheus monitoring.

As a side effect, if one queries now all hosts on s1 on zarcillo,
they will not show up. CC @marostegui

We need to find a better structure, one where it classifies by
replica set (e.g. for schema changes) and another by functionality
(core, vs. labs vs. test).

One option would be to rename the sections on test "s1" and "s4"
(but keep them part of the test group).
Another to treat specially test-s1 and test-s4. TBD.

For now, this patch shows test host back on prometheus, but not
as part of core group.

Change-Id: I235a9ac4f63c1b85b4964ef266bcb9b5f2e679c6