Rename attribute to enable MFA
For now we're setting the MFA method as a user-specific attribute and don't
use the use selection feature (which didn't work as adventised, we can revisit
this in a later CAS release).

As such, rename the attribute to the more fitting 'mfa-method' and remove the
logic to select additional attributes from the Groovy script. (But keep the
existing attributes in the LDAP schema, we'll probably use them some time in
the future)

Change-Id: I2d03de45beb983e21971bf18facdf4379353dd02