Optimize archiva-gitfat-link script
This script was previously excuting one large sha1sum command
for every artifact in Archiva, and then parsing the output
to create symlinks.

This change modifies the behavior to prefer to use the content of the
.sha1 files that Archiva itself already creates. Artifact paths will
be looped over.  If a .sha1 file for an artifact exists, its content
will be used as the git-fat symlink name.  If no .sha1 sum file
exists, then sha1sum of the file will be computed and that will
be used as the git-fat symlink name.

The downside is that this script will no longer fail if Archiva's
sha1 sum happens to be different than the computed sha1 sum of an
artifact.  In practice this should never happen anyway.

This optimization reduces the runtime of this script from about 5+ minutes
to ~45 seconds.

Bug: T235668
Change-Id: I26a956458413e39d5c18426c02015fa580e86e50