contint: downgrade docker on Stretch to match Jessie
The Docker package for Stretch is 5:18.09.7~3-0~debian-stretch, it is
crippled with a performance issue that cause CI jobs to take 20% more
time than they used to on Jessie.

Some long running job ends up timing out and the overhaul load is slower
on multiple jobs.  A long debugging session ruled out the OS, kernel,
cloduvirt, etc.   Using the version of Docker we had on Jessie proven to
be the fix.

We will docker-ce to be rolledback in apt.wikimedia.org to match the
Jessie version. Namely:

docker-ce 18.06.2~ce~3-0~debian

Release: stretch-wikimedia
Component: thirdparty/ci

Bug: T236675
Change-Id: I9e5ade681ff0b52e063e8f09a76f8a8282101ec3