parsoid-php: on beta, add sudo privs for php-fpm restarts
In production, the $use_php parameter switches a parsoid
server to include all the MW and PHP related classes which
includes the temp. setup for php-fpm restarts, specifically
the sudo prives to be able to do that.

In deployment-prep, the $use_php parameter is not currently
set in Hiera and instead multiple roles are directly applied
on the instance which pull in Mediawiki.

This means the sudo privileges are missing resulting in errors
like T236275#5621087.

Ideally the labs instance should also be able to just set $use_php
in Hiera but that requires probably more refactoring and making
a new instance starting with only the exact same role as used in prod.

Since this is a temporary setup anyways i'm just doing it this way instead.

Bug: T236275
Change-Id: I2e444c856ba7f1a4b7b6988954d84ada3629503b