tools-static: Allow X-Forwarded-Proto: https header
Right now this is directly exposed and terminating TLS all by itself.
I'd like to put it behind project-proxy, which will terminate TLS and set the
X-Forwarded-Proto header before passing it along to tools-static.

First let's get it accepting the X-Forwarded-Proto header as an indicator we
don't need to redirect. To do this we need a logical AND, and nginx doesn't
make that particularly nice.

Credit for this particular hack goes to:
https://www.digitalocean.com/community/questions/conditional-within-an-if-in-nginx?comment=42634
Bryan has pointed out that we do basically this trick in dynamicproxy.conf to
only enforce HTTPS on GET/HEAD requests (where redirecting is well handled).

Bug: T236952
Change-Id: I20fe46188d2d79503d051ebd7f778742bafd6450