toolforge: new k8s: rename hiera keys for consistency
Try to be consistent in the naming of related hiera keys.
This patch changes the following hiera keys:

* profile::toolforge::k8s_masters_hosts to profile::toolforge::k8s::control_nodes
* profile::toolforge::k8s::etcd_hosts to profile::toolforge::k8s::etcd_nodes
* profile::toolforge::k8s::apiserver to profile::toolforge::k8s::apiserver_fqdn
* drop profile::toolforge::k8s::api_servers and use profile::toolforge::k8s::control_nodes instead
* introduce profile::toolforge::k8s::apiserver_port to hold the apiserver port
* therefore, simplify the haproxy template for the apiservers

The intention is that setting up a k8s cluster is a bit easier because hiera
keys can be reused between profiles and they share a common naming pattern that
make them easier to understand.

Bug: T214513
Change-Id: Iabd8e496c5985bbbf4dffd55b701d80861b7da19
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>